--- a/include/tvm/relay/expr.h
+++ b/include/tvm/relay/expr.h
@@ -107,6 +107,7 @@ class TupleNode : public ExprNode {
   tvm::Array<relay::Expr> fields;
 
   void VisitAttrs(tvm::AttrVisitor* v) {
+    v->Visit("en_id", &en_id);
     v->Visit("fields", &fields);
     v->Visit("span", &span);
     v->Visit("_checked_type_", &checked_type_);
@@ -315,6 +316,7 @@ class CallNode : public ExprNode {
   tvm::Array<Type> type_args;
 
   void VisitAttrs(tvm::AttrVisitor* v) {
+    v->Visit("en_id", &en_id);
     v->Visit("op", &op);
     v->Visit("args", &args);
     v->Visit("attrs", &attrs);