--- a/python/tvm/driver/tvmc/composite_target.py
+++ b/python/tvm/driver/tvmc/composite_target.py
@@ -29,6 +29,7 @@ from tvm.relay.op.contrib.cmsisnn import partition_for_cmsisnn
 from tvm.relay.op.contrib.ethosu import partition_for_ethosu
 from tvm.relay.op.contrib.bnns import partition_for_bnns
 from tvm.relay.op.contrib.vitis_ai import partition_for_vitis_ai
+from tvm.relay.op.contrib.mrvl import partition_for_mrvl
 
 
 from .common import TVMCException
@@ -75,6 +76,10 @@ REGISTERED_CODEGEN = {
     "vitis-ai": {
         "config_key": "relay.ext.vitis_ai.options",
         "pass_pipeline": partition_for_vitis_ai,
+    },
+    "mrvl": {
+        "config_key": "relay.ext.mrvl.options",
+        "pass_pipeline": partition_for_mrvl,
     },
 }