@ccjoechou 2021-12-15
Hello @jroesch,
After done more self-discovery from reviewing other BYOC code files and CMake setup files, I think now I am aware of that, in the Mrvl.cmake file, the source file lists for libtvm_runtime.so (e.g., RUNTIME_SRCS) and for libtvm.so (e.g., COMPILER_SRCS) need to be separated. I have all BYOC-Marvell C++ files in the RUNTIME_SRCS list.
I will spend time to make adjustments and then run local TVM Jenkinsfile build & tests before I can git push a fix.
Thanks,
- Joe