@ssdurako 2021-12-15
Wondering if someone can help out provide some insights and pointers related to the failing "rust" tests.
In addition, there is quite a bit of "noise" with the cudart cluttering useful messages, so would appreciate if that can be suppressed/filtered out. - Thank you

2021-12-15 18:46:35.799320: W tensorflow/stream_executor/platform/default/dso_loader.cc:60] Could not load dynamic library 'libcudart.so.11.0'; dlerror: libcudart.so.11.0: cannot open shared object file: No such file or directory; LD_LIBRARY_PATH: